Extract paper outlining module
