+++ frontend/js/MainScript.js
@@ -1,5 +1,6 @@
 import { bootGridFinium } from '../scripts.js';
+import { detectPaperContour, extractContourPoints } from './PaperOutlining.js';
 
-bootGridFinium();
+bootGridFinium({ detectPaperContour, extractContourPoints });
 
 console.log('GridFinium MainScript: booted');
